REFACTOR-#0000: remove code duplication for 'to_pickle_distributed'
